Port script to work with Python 3
Use the python 3 interpreter.

Remove explicit gevent dependency from requirements.txt, and
rely on the package shipped with alpine linux,
to avoid having to build it with gcc.

Decode some bytes back into strings, although decoding with utf8 might
be incorrect, but i'm too lazy to check the DNS spec.

Add handling of additional exception types, especially when a host or
address is not found.